+++ apps/api.nameai.io/nameai/data/custom_dictionary.txt
