+++ apps/api.nameai.io/nameai/data/should_be_tokenized.txt
