+++ apps/api.nameai.io/nameai/data/tests/unicode14.txt
