+++ apps/api.nameai.io/nameai/data/tests/unicode15.txt
