+++ apps/api.nameai.io/nameai/data/tests/unicode_numerics.txt
