+++ apps/api.nameai.io/nameai/data/tests/unigram_freq.csv
