+++ apps/api.nameai.io/nameai/generate_static_data.py
@@ -1,0 +1,5 @@
+from nameai.static_property import generate_static_data
+
+
+if __name__ == '__main__':
+    generate_static_data()
