--- apps/api.nameai.io/namerank/generate_static_data.py
@@ -1,5 +1,0 @@
-from namerank.static_property import generate_static_data
-
-
-if __name__ == '__main__':
-    generate_static_data()
